# backend/openapi-gateway.json
{
  "components": {
    "schemas": {
      "AccidentPoint": {
        "properties": {
          "accident_number": {
            "title": "Accident Number",
            "type": "string"
          },
          "aircraft_category": {
            "title": "Aircraft Category",
            "type": "string"
          },
          "country": {
            "title": "Country",
            "type": "string"
          },
          "event_date": {
            "title": "Event Date",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "injury_severity": {
            "title": "Injury Severity",
            "type": "string"
          },
          "investigation_type": {
            "title": "Investigation Type",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "make": {
            "title": "Make",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "latitude",
          "longitude",
          "event_date",
          "location",
          "country",
          "injury_severity",
          "aircraft_category",
          "make",
          "model",
          "investigation_type",
          "accident_number"
        ],
        "title": "AccidentPoint",
        "type": "object"
      },
      "AccidentYearResponse": {
        "properties": {
          "accidents": {
            "items": {
              "$ref": "#/components/schemas/AccidentPoint"
            },
            "title": "Accidents",
            "type": "array"
          },
          "total_count": {
            "minimum": 0.0,
            "title": "Total Count",
            "type": "integer"
          },
          "unmapped_count": {
            "minimum": 0.0,
            "title": "Unmapped Count",
            "type": "integer"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "total_count",
          "unmapped_count",
          "accidents"
        ],
        "title": "AccidentYearResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "YearsResponse": {
        "properties": {
          "years": {
            "items": {
              "type": "integer"
            },
            "title": "Years",
            "type": "array"
          }
        },
        "required": [
          "years"
        ],
        "title": "YearsResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Aircraft Accident Map API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/accidents": {
      "get": {
        "operationId": "accidents_accidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": true,
            "schema": {
              "maximum": 9999,
              "minimum": 1000,
              "title": "Year",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccidentYearResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Accidents"
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/years": {
      "get": {
        "operationId": "years_years_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YearsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Years"
      }
    }
  },
  "x-google-allow": "configured",
  "x-google-backend": {
    "address": "https://backend.example.com",
    "path_translation": "APPEND_PATH_TO_ADDRESS"
  }
}